% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{convert.stellarhash}
\alias{convert.stellarhash}
\title{convert hash as presented on STELLAR network (base64 encoded) to standard hexadecimal value}
\usage{
convert.stellarhash(data)
}
\arguments{
\item{data}{base64 encoded hash}
}
\value{
hex hexadecimal hash
}
\description{
convert hash as presented on STELLAR network (base64 encoded) to standard hexadecimal value
}
\examples{
convert.stellarhash("KMVvhSYRAquk3lPpzljU4SytQSawsTz1aeB+PoKFaf0=")
}
