% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{convert_stellarHash}
\alias{convert_stellarHash}
\title{Convert hash on STELLAR network (base64 encoded) to standard hexadecimal value}
\usage{
convert_stellarHash(data)
}
\arguments{
\item{data}{base64 encoded hash}
}
\value{
hex hexadecimal hash
}
\description{
Convert hash on STELLAR network (base64 encoded) to standard hexadecimal value
}
\examples{
convert_stellarHash("KMVvhSYRAquk3lPpzljU4SytQSawsTz1aeB+PoKFaf0=")
}
