% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create.hash.file}
\alias{create.hash.file}
\title{create hash of a file}
\usage{
create.hash.file(data)
}
\arguments{
\item{data}{filename (and path, if outside working directory) of a file}
}
\value{
hash
}
\description{
create hash of a file
}
\examples{
create.hash.file("test.rds")
}
