% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create.hash.object}
\alias{create.hash.object}
\title{create hash of an object}
\usage{
create.hash.object(data)
}
\arguments{
\item{data}{any dataset or object}
}
\value{
hash
}
\description{
create hash of an object
}
\examples{
create.hash.object(data)
}
