% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create.tts.file}
\alias{create.tts.file}
\title{create trusted timestamp of a file}
\usage{
create.tts.file(data)
}
\arguments{
\item{data}{filename (and path, if outside working directory) of a file}
}
\value{
url
}
\description{
create trusted timestamp of a file
}
\examples{
create.tts.file("test.rds")
}
