% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create.tts.object}
\alias{create.tts.object}
\title{create trusted timestamp of an object}
\usage{
create.tts.object(data)
}
\arguments{
\item{data}{any dataset or object}
}
\value{
url
}
\description{
create trusted timestamp of an object
}
\examples{
create.tts.object(data)
}
