% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create_hashFile}
\alias{create_hashFile}
\title{Create hash of a file}
\usage{
create_hashFile(path)
}
\arguments{
\item{path}{filename (and path, if outside working directory) of a file}
}
\value{
hash
}
\description{
Create hash of a file
}
\examples{
create_hashFile("test.rds")
}
