% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create_hashObject}
\alias{create_hashObject}
\title{Create hash of an object/dataset}
\usage{
create_hashObject(data)
}
\arguments{
\item{data}{any dataset or object}
}
\value{
hash
}
\description{
Create hash of an object/dataset
}
\examples{
\donttest{
create_hashObject(data)
}
}
