% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create_ttsFile}
\alias{create_ttsFile}
\title{Create trusted timestamp of a file}
\usage{
create_ttsFile(path)
}
\arguments{
\item{path}{filename (and path, if outside working directory)}
}
\value{
url
}
\description{
Create trusted timestamp of a file
}
\examples{
create_ttsFile("test.rds")
}
