% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create_ttsObject}
\alias{create_ttsObject}
\title{Create trusted timestamp of an object/dataset}
\usage{
create_ttsObject(data)
}
\arguments{
\item{data}{any dataset or object}
}
\value{
url
}
\description{
Create trusted timestamp of an object/dataset
}
\examples{
create_ttsObject(data)
}
