% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get.hash}
\alias{get.hash}
\title{retrieve hash from STELLAR network}
\usage{
get.hash(url)
}
\arguments{
\item{url}{url}
}
\value{
hash
}
\description{
retrieve hash from STELLAR network
}
\examples{
get.hash("https://stellarapi.io/gethash/ea0ae0")
}
