% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get.timestamp}
\alias{get.timestamp}
\title{retrieve timestamp from STELLAR network}
\usage{
get.timestamp(url)
}
\arguments{
\item{url}{url}
}
\value{
GMT GMT-timestamp
}
\description{
retrieve timestamp from STELLAR network
}
\examples{
get.timestamp("https://stellarapi.io/gethash/ea0ae0")
}
