% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get.url.blockchaintransaction}
\alias{get.url.blockchaintransaction}
\title{retrieve url to blockchain transaction on STELLAR network}
\usage{
get.url.blockchaintransaction(url)
}
\arguments{
\item{url}{url}
}
\value{
url url to blockchain transaction
}
\description{
retrieve url to blockchain transaction on STELLAR network
}
\examples{
get.url.blockchaintransaction("https://stellarapi.io/gethash/ea0ae0")
}
