% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get_hash}
\alias{get_hash}
\title{Retrieve hash from STELLAR network}
\usage{
get_hash(url)
}
\arguments{
\item{url}{url}
}
\value{
hash
}
\description{
Retrieve hash from STELLAR network
}
\examples{
get_hash("https://stellarapi.io/gethash/ea0ae0")
}
