% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get_timestamp}
\alias{get_timestamp}
\title{Retrieve timestamp from STELLAR network}
\usage{
get_timestamp(url, proxy_ip = NULL, proxy_port = NULL)
}
\arguments{
\item{url}{url}

\item{proxy_ip}{if needed, provide proxy ip}

\item{proxy_port}{if needed, provide proxy port}
}
\value{
GMT GMT-timestamp
}
\description{
Retrieve timestamp from STELLAR network
}
\examples{
\donttest{
get_timestamp("https://stellarapi.io/gethash/ea0ae0")
}
}
