% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get_url_blockchaintransaction}
\alias{get_url_blockchaintransaction}
\title{Retrieve url of the transaction on STELLAR network}
\usage{
get_url_blockchaintransaction(url)
}
\arguments{
\item{url}{url}
}
\value{
url url of blockchain transaction
}
\description{
Retrieve url of the transaction on STELLAR network
}
\examples{
get_url_blockchaintransaction("https://stellarapi.io/gethash/ea0ae0")
}
