% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{GET}
\alias{digest}
\alias{fromJSON}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{digest}{\code{\link[digest]{digest}}}

  \item{httr}{\code{\link[httr]{GET}}}

  \item{jsonlite}{\code{\link[jsonlite]{fromJSON}}}
}}

