% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{validate.filehash}
\alias{validate.filehash}
\title{validate hash of a file (created on the fly) with hash in transaction on STELLAR network}
\usage{
validate.filehash(url, data)
}
\arguments{
\item{url}{url}

\item{data}{filename (and path, if outside working directory) of a file}
}
\value{
res result of validation
}
\description{
validate hash of a file (created on the fly) with hash in transaction on STELLAR network
}
\examples{
validate.filehash("https://stellarapi.io/gethash/ea0ae0", "test.rds")
}
