% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{validate.objecthash}
\alias{validate.objecthash}
\title{validate hash of an object (created on the fly) with hash in transaction on STELLAR network}
\usage{
validate.objecthash(url, data)
}
\arguments{
\item{url}{url}

\item{data}{any data/object}
}
\value{
res result of validation
}
\description{
validate hash of an object (created on the fly) with hash in transaction on STELLAR network
}
\examples{
validate.objecthash("https://stellarapi.io/gethash/ea0ae0", data)
}
