% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{validate_hashFile}
\alias{validate_hashFile}
\title{Validate hash of a file (created on the fly) with hash on STELLAR network}
\usage{
validate_hashFile(url, path)
}
\arguments{
\item{url}{url}

\item{path}{filename (and path, if outside working directory)}
}
\value{
res result of validation
}
\description{
Validate hash of a file (created on the fly) with hash on STELLAR network
}
\examples{
validate_hashFile("https://stellarapi.io/gethash/ea0ae0", "test.rds")
}
