% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create_ttsFile}
\alias{create_ttsFile}
\title{Create trusted timestamp of a file}
\usage{
create_ttsFile(path, proxy_ip = NULL, proxy_port = NULL)
}
\arguments{
\item{path}{filename (and path, if outside working directory)}

\item{proxy_ip}{if needed, provide proxy ip}

\item{proxy_port}{if needed, provide proxy port}
}
\value{
url
}
\description{
Create trusted timestamp of a file
}
\examples{
\donttest{
create_ttsFile("test.rds")
}
}
