% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{create_ttsObject}
\alias{create_ttsObject}
\title{Create trusted timestamp of an object/dataset}
\usage{
create_ttsObject(data, proxy_ip = NULL, proxy_port = NULL)
}
\arguments{
\item{data}{any dataset or object}

\item{proxy_ip}{ip if needed, provide proxy ip}

\item{proxy_port}{if needed, provide proxy port}
}
\value{
url
}
\description{
Create trusted timestamp of an object/dataset
}
\examples{
\dontrun{
create_ttsObject(data)
}
}
