% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get_hash}
\alias{get_hash}
\title{Retrieve hash from STELLAR network}
\usage{
get_hash(url, proxy_ip = NULL, proxy_port = NULL)
}
\arguments{
\item{url}{url}

\item{proxy_ip}{if needed, provide proxy ip}

\item{proxy_port}{if needed, provide proxy port}
}
\value{
hash
}
\description{
Retrieve hash from STELLAR network
}
\examples{
\donttest{
get_hash("https://horizon.stellar.org/transactions/ea0ae0etc")
}
}
