% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get_timestamp}
\alias{get_timestamp}
\title{Retrieve timestamp from STELLAR network}
\usage{
get_timestamp(url)
}
\arguments{
\item{url}{url}
}
\value{
GMT GMT-timestamp
}
\description{
Retrieve timestamp from STELLAR network
}
\examples{
get_timestamp("https://stellarapi.io/gethash/ea0ae0")
}
