% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{get_url_blockchaintransaction}
\alias{get_url_blockchaintransaction}
\title{Get url of the transaction on STELLAR network (stellarchain.io (non-json))}
\usage{
get_url_blockchaintransaction(url)
}
\arguments{
\item{url}{url}
}
\value{
url url of blockchain transaction
}
\description{
Get url of the transaction on STELLAR network (stellarchain.io (non-json))
}
\examples{
\donttest{
get_url_blockchaintransaction("https://horizon.stellar.org/transactions/ea0ae0etc")
}
}
