% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TTS-package.R
\docType{data}
\name{PC}
\alias{PC}
\title{Dataset obtained from polycarbonate (polymer) tests using Dynamic
Mechanical Analysis (DMA)}
\format{This data frame is composed of the following columns:
\describe{
\item{\code{log10.frequency}}{It accounts for seven different frequencies
           (rad/s) in logarithmic scale for each temperature (overall 49).}
\item{\code{log10.module}}{It accounts for seven different storage modulus,
           E' (Pa), in base-ten logarithmic scale for each temperature (overall 49).}
\item{\code{temperature}}{Seven different temperaturas: 147, 148, 149, 150,
           151, 152, 153 degrees celsius, each one with the corresponding seven values of
           frequency and storage modulus (overall 49).}
}}
\source{
Naya, S., Meneses A,. Tarrio-Saavedra, J,. Artiaga R.,
Lopez-Beceiro, J. and Gracia-Fernandez C. (2013) New method for estimating
shift factors in time-temperatura superposition models. Journal of Thermal
Analysis and Calorimetry. ISSN 1388-6150. DOI 10.1007/s10973-013-3193-1.
}
\description{
PC contains 49 rows and 3 columns.
}
\details{
The dataset corresponds to the storage modulus viscoelastic property of
different specimens of polycarbonate (PC) and obtained by DMA using TA
Instruments Q800 (Naya et al., 2013).
}
\examples{
data(PC)
}
\keyword{datasets}

