% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{validate_hashFile}
\alias{validate_hashFile}
\title{Validate hash of a file (created on the fly) with hash on STELLAR network
p.s. stellar transactions take between 5-7 seconds. If you validate to soon after creating a timestamp, it will fail...}
\usage{
validate_hashFile(url, path, proxy_ip = NULL, proxy_port = NULL)
}
\arguments{
\item{url}{url}

\item{path}{filename (and path, if outside working directory)}

\item{proxy_ip}{if needed, provide proxy ip}

\item{proxy_port}{if needed, provide proxy port}
}
\value{
res result of validation
}
\description{
Validate hash of a file (created on the fly) with hash on STELLAR network
p.s. stellar transactions take between 5-7 seconds. If you validate to soon after creating a timestamp, it will fail...
}
\examples{
\donttest{
validate_hashFile("https://stellarapi.io/gethash/ea0ae0", "test.rds")
}
}
