% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{validate_hashObject}
\alias{validate_hashObject}
\title{Validate hash of object/dataset (created on the fly) with hash on STELLAR network}
\usage{
validate_hashObject(url, data)
}
\arguments{
\item{url}{url}

\item{data}{any dataset or object}
}
\value{
res result of validation
}
\description{
Validate hash of object/dataset (created on the fly) with hash on STELLAR network
}
\examples{
validate_hashObject("https://stellarapi.io/gethash/ea0ae0", data)
}
