% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts.r
\name{validate_hashObject}
\alias{validate_hashObject}
\title{Validate hash of an object/dataset (created on the fly) with hash on STELLAR network
p.s. stellar transactions take between 5-7 seconds. If you validate to soon after creating a timestamp, it will fail...}
\usage{
validate_hashObject(url, data, proxy_ip = NULL, proxy_port = NULL)
}
\arguments{
\item{url}{url}

\item{data}{any dataset or object}

\item{proxy_ip}{if needed, provide proxy ip}

\item{proxy_port}{if needed, provide proxy port}
}
\value{
res result of validation
}
\description{
Validate hash of an object/dataset (created on the fly) with hash on STELLAR network
p.s. stellar transactions take between 5-7 seconds. If you validate to soon after creating a timestamp, it will fail...
}
\examples{
\donttest{
validate_hashObject("https://horizon.stellar.org/transactions/ea0ae0etc", data)
}
}
